Using Prepocessor Define to select MDS or Client for Message

+++ trunk/src/mds/namespacemodule.cc
@@ -32,7 +32,7 @@ vector<FileMetaData> NameSpaceModule::listFolder (uint32_t clientId, string path
 
 			tempFileMetaData._path = ent->d_name;
 			tempFileMetaData._size = tempFileStat.st_size;
-			debug("name: %s size: %d\n",ent->d_name,tempFileStat.st_size);
+			debug("name: %s size: %d\n",ent->d_name,(int)tempFileStat.st_size);
 			folderData.push_back(tempFileMetaData);
 		}
 		closedir (dir);
